+++ packages/eth-block-tracker/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- **BREAKING:** Use `InternalProvider` instead of `SafeEventEmitterProvider` ([#6796](https://github.com/MetaMask/core/pull/6796))
+  - The block tracker expects a provider with an `InternalProvider` instead of a `SafeEventEmitterProvider`.
+
 ## [14.0.0]
 
 ### Changed